+++ PCA9553.cpp
@@ -136,10 +136,9 @@ bool PCA9553::setLEDSource(uint8_t led, uint8_t source)
   if (led >= _channelCount) return false;
   if (source > 3) return false;
 
-  uint8_t val = source << (led * 2);
   uint8_t ledSelect = readReg(PCA9553_LS0);
   ledSelect &= ~(0x03 << (led * 2));
-  ledSelect |= val;
+  ledSelect |= (source << (led * 2));
 
   writeReg(PCA9553_LS0, ledSelect);
   return true;