+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - add **uint8_t getAddress()**
 - add **void digitalWrite(led, value)**
 - add **uint8_t digitalRead(led)**
+- add example
 - update readme.md
 - internal refactor