+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [0.2.1] - 2023-09-25
+- refactor API, begin();
+- update readme.md;
+- update examples
+
+----
+
+## [0.2.1] - 2023-09-25
 - add Wire1 support for ESP32
 - update readme.md
-
 
 ## [0.2.0] - 2023-07-17
 - refactor interface (breaking)